Fixed unsigned/signed warn

+++ d3d11game_xdk_dr/DeviceResources.cpp
@@ -260,8 +260,8 @@ void DeviceResources::CreateWindowSizeDependentResources()
         );
 
     m_outputSize.left = m_outputSize.top = 0;
-    m_outputSize.right = backBufferWidth;
-    m_outputSize.bottom = backBufferHeight;
+    m_outputSize.right = static_cast<LONG>(backBufferWidth);
+    m_outputSize.bottom = static_cast<LONG>(backBufferHeight);
 }
 
 // Prepare the render target for rendering.
